appsrc/testDriver.cpp:589:40: warning: taking the absolute value of unsigned type 'uint32' (aka 'unsigned int') has no effect [-Wabsolute-value]
  589 |         const uint64_t rand8432 = 1 + (labs(blindIncrFreq - tokenRandomizer)^(blindIncrFreq*tokenRandomizer));
      |                                        ^
appsrc/testDriver.cpp:589:40: note: remove the call to 'labs' since unsigned values cannot be negative
  589 |         const uint64_t rand8432 = 1 + (labs(blindIncrFreq - tokenRandomizer)^(blindIncrFreq*tokenRandomizer));
      |                                        ^~~~

+++ holdem/appsrc/testDriver.cpp
@@ -577,7 +577,7 @@ std::ifstream loadFile;
 
         const uint32 NUM_OPPONENTS = 8;
         const uint32 rand765 = 1 + ((blindIncrFreq + tokenRandomizer)^(blindIncrFreq*tokenRandomizer));
-        const uint64_t rand8432 = 1 + (labs(blindIncrFreq - tokenRandomizer)^(blindIncrFreq*tokenRandomizer));
+        const uint64_t rand8432 = 1 + (labs(static_cast<int32_t>(blindIncrFreq) - static_cast<int32_t>(tokenRandomizer))^(blindIncrFreq*tokenRandomizer));
         const uint32 rand8 = rand8432%8;
         const uint32 rand432 = static_cast<uint32>(rand8432/8);
         const uint32 randSeed = rand8 + (rand765%(7*6*5))*8 + (rand432%(4*3*2))*(8*7*6*5);
